factored out container structs for common sharing among other workers; new package named 'containermessage'
